requirements.txt
Django==4.1.7
pre-commit==3.1.1
djangorestframework==3.14.0